examples/co/requirements.txt
Django==1.6
south==1.0
selenium==2.44
pyrg==0.2.6
django-compressor==1.4
fabric
django-registration==1.0
django-widget-tweaks==1.3
django-taggit==0.12.2
celery==3.1.11
django-celery==3.1.10
redis==2.10.1
django-grappelli==2.5.3 
django-mptt==0.5.5
pytils==0.2.3
Pillow==2.5.1
sorl-thumbnail==11.12
django-image-cropping==1.0.0
easy-thumbnails==2.1
django-resized
django-ckeditor-updated==4.4.4
tornado==4.0.1
sockjs-tornado==1.0.1
-e git+https://github.com/peterbe/django-sockjs-tornado#egg=django-sockjs-tornado
djangoajax==2.2.12
-e git+https://github.com/zmathew/django-ajax-upload-widget.git#egg=ajax_upload
django-social-auth==0.7.28
django-rosetta==0.7.4
ffvideo
django-modeltranslation
rq==0.3.7
django-tastypie==0.9.12
django-wysiwyg-redactor
